Add viewing storage unit
Added ability to view storage unit and removed unused storage unit modals now that they have been moved.
