+++ api/src/routes/place.routes.ts
@@ -15,7 +15,7 @@ placeRoutes.get('/:placeId/object_instance',
   (request, response) => placeController.getPlaceObjects(request, response));
 placeRoutes.get('/:slug',
   (request, response) => placeController.getPlace(request, response));
-placeRoutes.get('/:id',
+placeRoutes.get('/by_id/:id',
   (request, response) => placeController.getPlaceById(request, response));
 placeRoutes.post('/add_storage', (request, response) => 
   placeController.addStorage(request, response));
