+++ spa/src/components/modals/InfoModal.vue
@@ -34,9 +34,7 @@
                   </a>
                 </td>
                 <td width=200 align=center>
-                  <a href="#" @click="openStorageModal">
-                    <img src="/assets/img/icon_myinfo.jpg" border=0><br><b>Storage Units</b>
-                  </a>
+                  <img src="/assets/img/icon_myinfo.jpg" border=0><br><b>Storage Units</b>
                 </td>
                 <td width=200 align=center>
                   <a href="#" @click="openAvatarModal">
@@ -61,7 +59,6 @@ import Vue from "vue";
 
 import Modal from './Modal.vue';
 import ModalMixin from './mixins/ModalMixin';
-import StorageModal from "./StorageModal.vue";
 import AvatarModal from "./AvatarModal.vue";
 import AccountModal from "./AccountModal.vue";
 import ModalService from "./services/ModalService.vue";
@@ -80,9 +77,6 @@ export default Vue.extend({
     openPersonalInfoModal(): void {
       ModalService.open(PersonalInfoModalVue);
     },
-    openStorageModal(): void {
-      ModalService.open(StorageModal);
-    }
   },
   mixins: [ModalMixin],
 });
