+++ spa/src/routes.ts
@@ -29,6 +29,7 @@ import HomeUpdatePage from "@/pages/home/HomeUpdatePage.vue";
 import HomeUpdateHomePage from "@/pages/home/HomeUpdateHomePage.vue";
 
 import UpdateStorageArea from "@/components/storage/Update.vue";
+import StorageUnit from "@/components/storage/Unit.vue";
 
 import MessageBoard from "@/pages/MessageBoard.vue";
 import Inbox from "@/pages/Inbox.vue";
@@ -579,6 +580,14 @@ export default [
             default: UpdateStorageArea,
         },
         name: "update-storage",
+        meta: { wrapper: false },
+    },
+    {
+        path: "/storage/unit/:id",
+        components: {
+            default: StorageUnit,
+        },
+        name: "storage-unit",
         meta: { wrapper: false },
     },
 ];
